+++ src/node-srv/browser/httpRPC.js
@@ -1,5 +1,5 @@
-class httpRPC {
-    constructor(httpOrs, host, port) {
+var httpRPC = (function () {
+    function httpRPC(httpOrs, host, port) {
         this.user = '';
         this.pswd = '';
         this.token = '';
@@ -9,14 +9,14 @@ class httpRPC {
         this.port = port;
         console.log(this.httpOrs, this.host, this.port);
     }
-    setUser(user, pswd) {
+    httpRPC.prototype.setUser = function (user, pswd) {
         this.user = user;
         this.pswd = pswd;
-    }
-    setToken(token) {
+    };
+    httpRPC.prototype.setToken = function (token) {
         this.token = token;
-    }
-    invoke(route, method, params) {
+    };
+    httpRPC.prototype.invoke = function (route, method, params) {
         if (!params)
             params = {};
         params.method = method;
@@ -26,9 +26,9 @@ class httpRPC {
         params.view = window.location.href;
         var str = JSON.stringify(params);
         var compressed = LZString.compressToEncodedURIComponent(str);
-        const THIZ = this;
+        var THIZ = this;
         return new Promise(function (resolve, reject) {
-            let url = THIZ.httpOrs + '://' + THIZ.host + (THIZ.port ? (':' + THIZ.port) : '') + '/' + route;
+            var url = THIZ.httpOrs + '://' + THIZ.host + (THIZ.port ? (':' + THIZ.port) : '') + '/' + route;
             url = url + '/?p=' + compressed;
             fetch(url, {
                 method: 'GET',
@@ -62,16 +62,16 @@ class httpRPC {
                 reject(method + ' ' + err);
             });
         });
-    }
-    setItem(key, val) {
+    };
+    httpRPC.prototype.setItem = function (key, val) {
         sessionStorage.setItem(key, val);
-    }
-    getItem(key) {
+    };
+    httpRPC.prototype.getItem = function (key) {
         return sessionStorage.getItem(key);
-    }
-    log(msg, level, className) {
+    };
+    httpRPC.prototype.log = function (msg, level, className) {
         var THIZ = this;
-        let p = {
+        var p = {
             msg: msg,
             page: window.location.pathname,
             level: level,
@@ -95,5 +95,6 @@ class httpRPC {
             console.log(className, level, msg);
         else
             console.log(msg);
-    }
-}
+    };
+    return httpRPC;
+}());