+++ src/node-srv/lib/SrvRPC.js
@@ -1,20 +1,21 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const fetch = require('make-fetch-happen');
 const bunyan = require('bunyan');
 const bformat = require('bunyan-format2');
 const formatOut = bformat({ outputMode: 'short' });
-const log = bunyan.createLogger({ src: true, stream: formatOut, name: "invoke" });
 const btoa = function (str) { return Buffer.from(str).toString('base64'); };
+const fetch = require('make-fetch-happen');
+const LZString = require('lz-string');
 class HttpRPC {
     constructor(httpOrs, host, port) {
+        this._log = bunyan.createLogger({ src: true, stream: formatOut, name: this.constructor.name });
         this.user = '';
         this.pswd = '';
         this.token = '';
         this.httpOrs = httpOrs;
         this.host = host;
         this.port = port;
-        log.info(this.httpOrs, this.host, this.port);
+        this._log.info(this.httpOrs, this.host, this.port);
     }
     setUser(user, pswd) {
         this.user = user;
@@ -23,21 +24,19 @@ class HttpRPC {
     setToken(token) {
         this.token = token;
     }
-    invoke(route, ent, method, params) {
+    invoke(route, method, params) {
         if (!params)
             params = {};
-        params.ent = ent;
         params.method = method;
         params.user = btoa(this.user);
         params.pswd = btoa(this.pswd);
         params.token = btoa(this.token);
-        let query = Object.keys(params)
-            .map(k => encodeURIComponent(k) + '=' + encodeURIComponent(params[k]))
-            .join('&');
+        let str = JSON.stringify(params);
+        const compressed = LZString.compressToEncodedURIComponent(str);
         const THIZ = this;
         return new Promise(function (resolve, reject) {
             let url = THIZ.httpOrs + '://' + THIZ.host + (THIZ.port ? (':' + THIZ.port) : '') + '/' + route;
-            url = url + '/?' + query;
+            url = url + '/?p=' + compressed;
             fetch(url, {
                 method: 'GET',
                 cache: 'default',
@@ -46,22 +45,24 @@ class HttpRPC {
                 keepalive: true
             })
                 .then(function (fullResp) {
-                const obj = fullResp.json();
-                if (!fullResp.ok)
-                    reject(obj);
-                else {
-                    return obj;
+                if (!fullResp.ok) {
+                    THIZ._log.warn('HTTP protocol error in RPC: ' + fullResp.status + fullResp);
+                    reject('HTTP protocol error in RPC: ' + fullResp.status + fullResp);
                 }
+                else
+                    return fullResp.text();
             })
-                .then(function (resp) {
-                if (resp.errorMessage) {
-                    reject(resp);
+                .then(function (str) {
+                const resp = JSON.parse(str);
+                if ((!resp) || resp.errorMessage) {
+                    THIZ._log.warn(method + ' ' + str);
+                    reject(method + ' ' + str);
                 }
                 resolve(resp.result);
             })
                 .catch(function (err) {
-                log.info('fetch err');
-                log.info(err);
+                THIZ._log.warn('fetch err');
+                THIZ._log.warn(err);
                 reject(err);
             });
         });