+++ src/node-srv/lib/SrvRPC.ts
@@ -1,13 +1,13 @@
 // All rights reserved by Cekvenich|INTUITION.DEV) |  Cekvenich, licensed under LGPL 3.0
 
-const fetch = require('make-fetch-happen')
 
 const bunyan = require('bunyan')
 const bformat = require('bunyan-format2')  
 const formatOut = bformat({ outputMode: 'short' })
-const log = bunyan.createLogger({src: true, stream: formatOut, name: "invoke"})
 
 const btoa = function(str){ return Buffer.from(str).toString('base64'); }
+const fetch = require('make-fetch-happen')
+const LZString = require('lz-string')
 
 // requires promise and fetch for ie11, you should require 'poly'
 export class HttpRPC {// 
@@ -15,6 +15,9 @@ export class HttpRPC {//
     httpOrs // protocol
     host
     port
+
+    _log = bunyan.createLogger({src: true, stream: formatOut, name: this.constructor.name })
+
   
     /**
      * 
@@ -33,7 +36,7 @@ export class HttpRPC {//
       this.host = host
       this.port = port
   
-      log.info(this.httpOrs, this.host, this.port)
+      this._log.info(this.httpOrs, this.host, this.port)
   
     }
     //apiPath=''
@@ -50,31 +53,27 @@ export class HttpRPC {//
   
     /**
      * @param route api apth, eg api
-     * @param ent  viewmodel name | page name | screen name | component name | calling url | ECSid 
      * @param method CRUD, insert, check, listAll, etc
      * @param params Object of name value pairs.
      */
-    invoke(route, ent, method, params):Promise<string> { // returns promise of results or err
+    invoke(route, method, params):Promise<string> { // returns promise of results or err
       //if array, return as array
       if(!params) params = {}
   
-      params.ent=ent
       params.method=method
       params.user = btoa(this.user)
       params.pswd = btoa(this.pswd)
       params.token = btoa(this.token)
   
-      let query = Object.keys(params)
-               .map(k => encodeURIComponent(k) + '=' + encodeURIComponent(params[k]))
-               .join('&')
-  
+      let str = JSON.stringify(params)
+      const compressed = LZString.compressToEncodedURIComponent(str)
+
       const THIZ = this
       return new Promise(function(resolve, reject) {
         //log.info(formData.get('method'))
         let url:string = THIZ.httpOrs+'://'+THIZ.host + (THIZ.port ? (':' + THIZ.port) : '') + '/'+route 
   
-        url = url + '/?' + query
-      
+        url = url + '/?p=' + compressed
         fetch(url, {
               method: 'GET',
               cache: 'default',
@@ -83,25 +82,29 @@ export class HttpRPC {//
               //credentials: 'include',
               keepalive: true
   
-            })//fetch
-            .then(function(fullResp) {
-              const obj = fullResp.json();
-              
-              if(!fullResp.ok) 
-                reject(obj)
-               else {
-                return obj
-              }
             })
-            .then(function(resp) {
-              if(resp.errorMessage) {
-                reject(resp)
-              }
-              resolve(resp.result)
+            .then(function(fullResp) {
+
+              if(!fullResp.ok) {
+                THIZ._log.warn('HTTP protocol error in RPC: ' + fullResp.status + fullResp)
+                reject('HTTP protocol error in RPC: ' + fullResp.status + fullResp)
+             }
+              else 
+                return fullResp.text()
+           })
+           .then(function(str) {     
+              const resp = JSON.parse(str)
+              
+              if((!resp) || resp.errorMessage) {
+                THIZ._log.warn(method +' '+ str)
+                reject(method +' '+ str)
+             }
+             resolve(resp.result)
+
             })//fetch
             .catch(function (err) {
-              log.info('fetch err')
-              log.info(err)
+              THIZ._log.warn('fetch err')
+              THIZ._log.warn(err)
               reject(err)
             })
         })//pro